Refactor: if문 플랫화, try문 삭제, 함수 파일 분리
